requirements.txt
cryptography==42.0.5
Django==5.0.2
django-allauth==0.61.1
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
python-dateutil==2.9.0.post0
ruff==0.3.2
